Back to origin data type

+++ src/WebApplication4/Models/FoodCalculator.cs
@@ -26,9 +26,9 @@ namespace WebApplication4.Models
 
         public int Grams { get; set; }
 
-        public float Calories { get; set; }
+        public int Calories { get; set; }
 
-        public float Lipid { get; set; }
+        public int Lipid { get; set; }
 
         [Required]
         [DisplayFormat(DataFormatString = "{0:MM/dd/yyyy}")]
